Moves remaining tags into new files
