+++ bookwyrm/templates/about/about.html
@@ -2,7 +2,7 @@
 {% load humanize %}
 {% load i18n %}
 {% load utilities %}
-{% load bookwyrm_tags %}
+{% load landing_page_tags %}
 {% load cache %}
 
 {% block title %}
