+++ bookwyrm/templates/book/book.html
@@ -1,6 +1,6 @@
 {% extends 'layout.html' %}
 {% load i18n %}
-{% load bookwyrm_tags %}
+{% load book_display_tags %}
 {% load humanize %}
 {% load utilities %}
 {% load static %}
