+++ bookwyrm/templates/book/file_links/links.html
@@ -1,5 +1,5 @@
 {% load i18n %}
-{% load bookwyrm_tags %}
+{% load book_display_tags %}
 {% load utilities %}
 
 {% get_book_file_links book as links %}
