+++ bookwyrm/templates/discover/small-book.html
@@ -1,4 +1,4 @@
-{% load bookwyrm_tags %}
+{% load landing_page_tags %}
 {% load utilities %}
 {% load i18n %}
 {% load status_display %}
