+++ bookwyrm/templates/feed/status.html
@@ -1,6 +1,6 @@
 {% extends 'feed/layout.html' %}
+{% load feed_page_tags %}
 {% load i18n %}
-{% load bookwyrm_tags %}
 
 {% block opengraph_images %}
