+++ bookwyrm/templates/feed/suggested_books.html
@@ -1,5 +1,5 @@
 {% load i18n %}
-{% load bookwyrm_tags %}
+{% load feed_page_tags %}
 
 {% suggested_books as suggested_books %}
 <section class="block">
