+++ bookwyrm/templates/groups/group.html
@@ -1,6 +1,5 @@
 {% extends 'groups/layout.html' %}
 {% load i18n %}
-{% load bookwyrm_tags %}
 {% load bookwyrm_group_tags %}
 {% load markdown %}
