+++ bookwyrm/templates/groups/members.html
@@ -1,7 +1,6 @@
 {% load i18n %}
 {% load utilities %}
 {% load humanize %}
-{% load bookwyrm_tags %}
 {% load bookwyrm_group_tags %}
 
 <h2 class="title is-5">Group Members</h2>
