+++ bookwyrm/templates/landing/landing.html
@@ -1,7 +1,7 @@
 {% extends 'landing/layout.html' %}
 {% load i18n %}
 {% load cache %}
-{% load bookwyrm_tags %}
+{% load landing_page_tags %}
 
 {% block panel %}
