+++ bookwyrm/templates/notifications/items/layout.html
@@ -1,4 +1,4 @@
-{% load bookwyrm_tags %}
+{% load notification_page_tags %}
 {% related_status notification as related_status %}
 <div class="notification {% if notification.id in unread %}has-background-primary{% endif %}">
     <div class="columns is-mobile {% if notification.id in unread %}has-text-white{% else %}has-text-grey{% endif %}">
