+++ bookwyrm/templates/snippets/create_status/layout.html
@@ -1,4 +1,3 @@
-{% load bookwyrm_tags %}
 {% load i18n %}
 {% load utilities %}
 {% load status_display %}
