+++ bookwyrm/templates/snippets/create_status/quotation.html
@@ -1,5 +1,4 @@
 {% extends "snippets/create_status/layout.html" %}
-{% load bookwyrm_tags %}
 {% load utilities %}
 {% load status_display %}
 {% load i18n %}
