+++ bookwyrm/templates/snippets/status/content_status.html
@@ -1,4 +1,4 @@
-{% load bookwyrm_tags %}
+{% load book_display_tags %}
 {% load rating_tags %}
 {% load markdown %}
 {% load i18n %}
