+++ bookwyrm/templates/snippets/status/generated_status.html
@@ -1,6 +1,6 @@
 {% spaceless %}
 
-{% load bookwyrm_tags %}
+{% load book_display_tags %}
 {% load markdown %}
 {% load i18n %}
 {% load cache %}
