+++ bookwyrm/templates/user/user_preview.html
@@ -1,7 +1,7 @@
 {% load i18n %}
 {% load humanize %}
 {% load utilities %}
-{% load bookwyrm_tags %}
+{% load user_page_tags %}
 
 <div class="media block">
     <div class="media-left">
