+++ bookwyrm/templatetags/landing_page_tags.py
@@ -3,30 +3,8 @@ from django import template
 from django.db.models import Avg, StdDev, Count, F, Q
 
 from bookwyrm import models
-from bookwyrm.views.feed import get_suggested_books
-
 
 register = template.Library()
-
-
-@register.filter(name="book_description")
-def get_book_description(book):
-    """use the work's text if the book doesn't have it"""
-    return book.description or book.parent_work.description
-
-
-@register.filter(name="load_subclass")
-def load_subclass(status):
-    """sometimes you didn't select_subclass"""
-    if hasattr(status, "quotation"):
-        return status.quotation
-    if hasattr(status, "review"):
-        return status.review
-    if hasattr(status, "comment"):
-        return status.comment
-    if hasattr(status, "generatednote"):
-        return status.generatednote
-    return status
 
 
 @register.simple_tag(takes_context=False)
@@ -81,14 +59,6 @@ def get_book_superlatives():
 
 
 @register.simple_tag(takes_context=False)
-def related_status(notification):
-    """for notifications"""
-    if not notification.related_status:
-        return None
-    return load_subclass(notification.related_status)
-
-
-@register.simple_tag(takes_context=False)
 def get_landing_books():
     """list of books for the landing page"""
     return list(
@@ -104,26 +74,3 @@ def get_landing_books():
             .order_by("-review__published_date")[:6]
         )
     )
-
-
-@register.simple_tag(takes_context=True)
-def mutuals_count(context, user):
-    """how many users that you follow, follow them"""
-    viewer = context["request"].user
-    if not viewer.is_authenticated:
-        return None
-    return user.followers.filter(followers=viewer).count()
-
-
-@register.simple_tag(takes_context=True)
-def suggested_books(context):
-    """get books for suggested books panel"""
-    # this happens here instead of in the view so that the template snippet can
-    # be cached in the template
-    return get_suggested_books(context["request"].user)
-
-
-@register.simple_tag(takes_context=False)
-def get_book_file_links(book):
-    """links for a book"""
-    return book.file_links.filter(domain__status="approved")
