+++ bookwyrm/templatetags/user_page_tags.py
@@ -1,0 +1,14 @@
+""" template filters """
+from django import template
+
+
+register = template.Library()
+
+
+@register.simple_tag(takes_context=True)
+def mutuals_count(context, user):
+    """how many users that you follow, follow them"""
+    viewer = context["request"].user
+    if not viewer.is_authenticated:
+        return None
+    return user.followers.filter(followers=viewer).count()
